docs: add module docstrings
