+++ docs/source/smarter-resources/account/models.rst
@@ -1,5 +1,5 @@
-Django ORM
-================
+Account Django ORM
+===================
 
 .. toctree::
    :maxdepth: 1
