+++ docs/source/smarter-resources/account/views/plugin.rst
@@ -5,3 +5,4 @@ Plugin Views
     :members:
     :undoc-members:
     :show-inheritance:
+    :no-index:
